+++ app/src/main/java/com/feitianzhu/huangliwo/common/base/activity/AbsActivity.java
@@ -13,6 +13,9 @@ import com.afollestad.materialdialogs.MaterialDialog;
 import com.feitianzhu.huangliwo.GlobalUtil;
 import com.feitianzhu.huangliwo.R;
 import com.gyf.immersionbar.ImmersionBar;
+import com.lxj.xpopup.XPopup;
+import com.lxj.xpopup.enums.PopupAnimation;
+import com.lxj.xpopup.impl.LoadingPopupView;
 import com.umeng.message.PushAgent;
 
 import butterknife.ButterKnife;
@@ -28,6 +31,7 @@ import butterknife.Unbinder;
 
 public abstract class AbsActivity extends AppCompatActivity {
     private MaterialDialog mDialog;
+    LoadingPopupView loadingPopup;
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
@@ -60,29 +64,37 @@ public abstract class AbsActivity extends AppCompatActivity {
 
 
     protected void showloadDialog(String title) {
-        mDialog = new MaterialDialog.Builder(this)
-                .content("加载中,请稍等")
-                .progress(true, 0)
-                .progressIndeterminateStyle(false)
-                .show();
-    }
-
-    protected void showloadDialogText(String title) {
-        mDialog = new MaterialDialog.Builder(this)
-                .content(title)
-                .progress(true, 0)
-                .progressIndeterminateStyle(false)
+        loadingPopup = (LoadingPopupView) new XPopup.Builder(this)
+                .hasShadowBg(false)
+                .popupAnimation(PopupAnimation.NoAnimation)
+                .asLoading()
+                .bindLayout(R.layout.layout_loading_view)
                 .show();
     }
 
     protected void goneloadDialog() {
-        if (null != mDialog && mDialog.isShowing()) mDialog.dismiss();
+        if (null != loadingPopup) {
+            loadingPopup.delayDismissWith(600, new Runnable() {
+                @Override
+                public void run() {
+                }
+            });
+        }
     }
+
+//    protected void showloadDialogText(String title) {
+//        mDialog = new MaterialDialog.Builder(this)
+//                .content(title)
+//                .progress(true, 0)
+//                .progressIndeterminateStyle(false)
+//                .show();
+//    }
+
 
     @Override
     protected void onDestroy() {
         super.onDestroy();
-        mDialog = null;
+        loadingPopup = null;
     }
 
     @Override
